requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# RL Libraries
gymnasium>=0.28.0
stable-baselines3>=2.0.0

# SUMO Integration
#sumolib>=1.19.0
#traci>=1.19.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Logging & Experiment Tracking
tensorboard>=2.13.0
wandb>=0.15.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
tqdm>=4.65.0
click>=8.1.0
python-dotenv>=1.0.0

# Data Processing
scikit-learn>=1.3.0
opencv-python-headless>=4.7.0

# Development
pytest>=7.3.0
pytest-cov>=4.1.0
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0

